requirements.txt
streamlit==1.31.1
plotly==5.19.0
lxml==5.1.0
openpyxl==3.1.2
python-docx==1.1.0